requirements.txt
streamlit==1.40.0
psycopg2-binary==2.9.11
pandas==2.2.3
numpy==1.26.4
plotly==5.24.1
altair==5.5.0
python-dotenv==1.1.0
sqlalchemy==2.0.35
pyarrow==17.0.0
